requirements.txt
alembic==0.9.1
appdirs==1.4.3
beautifulsoup4==4.5.3
click==6.7
coverage==4.3.4
cssselect==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.1
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.9.6
lxml==3.7.3
Mako==1.0.6
MarkupSafe==1.0
newspaper3k==0.1.9
nltk==3.2.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
PyYAML==3.12
requests==2.13.0
requests-file==1.4.1
six==1.10.0
SQLAlchemy==1.1.9
tldextract==2.0.2
Werkzeug==0.12.1
